+++ sdk/search/azure-search-documents/azure/search/documents/_internal/aio/_search_indexing_buffered_sender_async.py
@@ -141,12 +141,12 @@ class SearchIndexingBufferedSender(SearchIndexingBufferedSenderBase, HeadersMixi
                 try:
                     action = next(x for x in actions if x.additional_properties.get(self._index_key) == result.key)
                     if result.succeeded:
-                        await self._succeed_callback(action)
+                        await self._callback_succeed(action)
                     elif is_retryable_status_code(result.status_code):
                         await self._retry_action(action)
                         has_error = True
                     else:
-                        await self._fail_callback(action)
+                        await self._callback_fail(action)
                         has_error = True
                 except StopIteration:
                     pass
@@ -192,7 +192,7 @@ class SearchIndexingBufferedSender(SearchIndexingBufferedSenderBase, HeadersMixi
         :type documents: List[dict]
         """
         actions = await self._index_documents_batch.add_upload_actions(documents)
-        await self._new_action_callback(actions)
+        await self._callback_new(actions)
         await self._process_if_needed()
 
     @distributed_trace_async
@@ -203,7 +203,7 @@ class SearchIndexingBufferedSender(SearchIndexingBufferedSenderBase, HeadersMixi
         :type documents: List[dict]
         """
         actions = await self._index_documents_batch.add_delete_actions(documents)
-        await self._new_action_callback(actions)
+        await self._callback_new(actions)
         await self._process_if_needed()
 
     @distributed_trace_async
@@ -214,7 +214,7 @@ class SearchIndexingBufferedSender(SearchIndexingBufferedSenderBase, HeadersMixi
         :type documents: List[dict]
         """
         actions = await self._index_documents_batch.add_merge_actions(documents)
-        await self._new_action_callback(actions)
+        await self._callback_new(actions)
         await self._process_if_needed()
 
     @distributed_trace_async
@@ -225,7 +225,7 @@ class SearchIndexingBufferedSender(SearchIndexingBufferedSenderBase, HeadersMixi
         :type documents: List[dict]
         """
         actions = await self._index_documents_batch.add_merge_or_upload_actions(documents)
-        await self._new_action_callback(actions)
+        await self._callback_new(actions)
         await self._process_if_needed()
 
     @distributed_trace_async
@@ -296,7 +296,7 @@ class SearchIndexingBufferedSender(SearchIndexingBufferedSenderBase, HeadersMixi
     async def _retry_action(self, action):
         # type: (IndexAction) -> None
         if not self._index_key:
-            await self._fail_callback(action)
+            await self._callback_fail(action)
             return
         key = action.additional_properties.get(self._index_key)
         counter = self._retry_counter.get(key)
@@ -309,24 +309,24 @@ class SearchIndexingBufferedSender(SearchIndexingBufferedSenderBase, HeadersMixi
             self._retry_counter[key] = counter + 1
             await self._index_documents_batch.enqueue_action(action)
         else:
-            await self._fail_callback(action)
+            await self._callback_fail(action)
 
-    async def _succeed_callback(self, action):
+    async def _callback_succeed(self, action):
         # type: (IndexAction) -> None
-        if self._remove_callback:
-            await self._remove_callback(action)
-        if self._progress_callback:
-            await self._progress_callback(action)
+        if self._on_remove:
+            await self._on_remove(action)
+        if self._on_progress:
+            await self._on_progress(action)
 
-    async def _fail_callback(self, action):
+    async def _callback_fail(self, action):
         # type: (IndexAction) -> None
-        if self._remove_callback:
-            await self._remove_callback(action)
-        if self._error_callback:
-            await self._error_callback(action)
+        if self._on_remove:
+            await self._on_remove(action)
+        if self._on_error:
+            await self._on_error(action)
 
-    async def _new_action_callback(self, actions):
+    async def _callback_new(self, actions):
         # type: (List[IndexAction]) -> None
-        if self._new_callback:
+        if self._on_new:
             for action in actions:
-                await self._new_callback(action)
+                await self._on_new(action)